more cleaning up

+++ src/app/pdf-tree/pdf-tree.component.ts
@@ -1,4 +1,3 @@
-import { SelectionModel } from "@angular/cdk/collections";
 import { FlatTreeControl } from "@angular/cdk/tree";
 import { Component, ElementRef, ViewChild } from "@angular/core";
 import { MatTreeFlatDataSource, MatTreeFlattener } from "@angular/material/tree";
@@ -10,36 +9,23 @@ import { FlatNode, Node, PdfBuilder } from "../services/pdf-builder.service";
     styleUrls: ["pdf-tree.component.scss"],
 })
 export class PdfTreeComponent {
-    /** Map from flat node to nested node. This helps us finding the nested node to be modified */
-    public flatNodeMap = new Map<FlatNode, Node>();
-
-    /** Map from nested node to flattened node. This helps us to keep the same object for selection */
-    public nestedNodeMap = new Map<Node, FlatNode>();
-
-    /** A selected parent node to be inserted */
-    public selectedParent: FlatNode | null = null;
-
-    /** The new item's name */
-    public newItemName = "";
 
     public treeControl: FlatTreeControl<FlatNode>;
-
-    public treeFlattener: MatTreeFlattener<Node, FlatNode>;
-
     public dataSource: MatTreeFlatDataSource<Node, FlatNode>;
 
-    /** The selection for checklist */
-    public checklistSelection = new SelectionModel<FlatNode>(true /* multiple */);
+    public dragNodeExpandOverNode: FlatNode;
+    public dragNodeExpandOverArea: "above" | "below" | "center";
 
-    /* Drag and drop */
-    public dragNode: any;
-    public dragNodeExpandOverWaitTimeMs = 300;
-    public dragNodeExpandOverNode: any;
-    public dragNodeExpandOverTime: number;
-    public dragNodeExpandOverArea: string;
+    private flatNodeMap = new Map<FlatNode, Node>();
+    private nestedNodeMap = new Map<Node, FlatNode>();
+    private treeFlattener: MatTreeFlattener<Node, FlatNode>;
+
+    private dragNode: FlatNode;
+    private dragNodeExpandOverWaitTimeMs = 300;
+    private dragNodeExpandOverTime: number;
 
     @ViewChild("emptyItem", { static: true })
-    public emptyItem: ElementRef;
+    private emptyItem: ElementRef;
 
     constructor(
         private database: PdfBuilder,
@@ -62,9 +48,7 @@ export class PdfTreeComponent {
 
     public hasChild = (_: number, nodeData: FlatNode) => nodeData.expandable;
 
-    /**
-     * Transformer to convert nested node to flat node. Record the nodes in maps for later use.
-     */
+    // Transformer to convert nested node to flat node. Record the nodes in maps for later use.
     public transformer = (node: Node, level: number) => {
         const existingNode = this.nestedNodeMap.get(node);
         const flatNode = existingNode && existingNode.item === node.item
@@ -78,36 +62,13 @@ export class PdfTreeComponent {
         return flatNode;
     }
 
-    /** Whether all the descendants of the node are selected */
-    public descendantsAllSelected(node: FlatNode): boolean {
-        const descendants = this.treeControl.getDescendants(node);
-        return descendants.every((child) => this.checklistSelection.isSelected(child));
-    }
-
-    /** Whether part of the descendants are selected */
-    public descendantsPartiallySelected(node: FlatNode): boolean {
-        const descendants = this.treeControl.getDescendants(node);
-        const result = descendants.some((child) => this.checklistSelection.isSelected(child));
-        return result && !this.descendantsAllSelected(node);
-    }
-
-    /** Toggle the to-do item selection. Select/deselect all the descendants node */
-    public todoItemSelectionToggle(node: FlatNode): void {
-        this.checklistSelection.toggle(node);
-        const descendants = this.treeControl.getDescendants(node);
-        this.checklistSelection.isSelected(node)
-            ? this.checklistSelection.select(...descendants)
-            : this.checklistSelection.deselect(...descendants);
-    }
-
-    /** Select the category so we can insert the new item. */
     public addNewItem(node: FlatNode) {
         const parentNode = this.flatNodeMap.get(node);
         // this.database.insertItem(parentNode, "");
         this.treeControl.expand(node);
     }
 
-    public handleDragStart(event, node) {
+    public handleDragStart(event: DragEventInit, node: FlatNode) {
         // Required by Firefox (https://stackoverflow.com/questions/19055264/why-doesnt-html5-drag-and-drop-work-in-firefox)
         event.dataTransfer.setData("foo", "bar");
         event.dataTransfer.setDragImage(this.emptyItem.nativeElement, 0, 0);
@@ -115,7 +76,7 @@ export class PdfTreeComponent {
         this.treeControl.collapse(node);
     }
 
-    public handleDragOver(event, node) {
+    public handleDragOver(event: DragEvent, node: FlatNode) {
         event.preventDefault();
 
         // Handle node expand
@@ -131,8 +92,9 @@ export class PdfTreeComponent {
         }
 
         // Handle drag area
-        const percentageX = event.offsetX / event.target.clientWidth;
-        const percentageY = event.offsetY / event.target.clientHeight;
+        const percentageX = event.offsetX / (event.target as HTMLElement).clientWidth;
+        const percentageY = event.offsetY / (event.target as HTMLElement).clientHeight;
+
         if (percentageY < 0.25) {
             this.dragNodeExpandOverArea = "above";
         } else if (percentageY > 0.75) {
@@ -142,26 +104,37 @@ export class PdfTreeComponent {
         }
     }
 
-    public handleDrop(event, node) {
+    public handleDrop(event: DragEvent, node: FlatNode) {
         event.preventDefault();
         if (node !== this.dragNode) {
             let newItem: Node;
-            if (this.dragNodeExpandOverArea === "above") {
-                newItem = this.database.copyPasteItemAbove(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
-            } else if (this.dragNodeExpandOverArea === "below") {
-                newItem = this.database.copyPasteItemBelow(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
-            } else {
-                newItem = this.database.copyPasteItem(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
+
+            switch (this.dragNodeExpandOverArea) {
+                case "above":
+                    newItem = this.database.copyPasteItemAbove(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
+                    break;
+                case "below":
+                    newItem = this.database.copyPasteItemBelow(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
+                    break;
+                case "center":
+                    newItem = this.database.copyPasteItem(this.flatNodeMap.get(this.dragNode), this.flatNodeMap.get(node));
+                    break;
+                default:
+                    console.error(`Invalid dragNodeExpandOverArea ${this.dragNodeExpandOverArea}`);
+                    break;
             }
-            this.database.deleteItem(this.flatNodeMap.get(this.dragNode));
-            this.treeControl.expandDescendants(this.nestedNodeMap.get(newItem));
+
+            if (newItem) {
+                this.database.deleteItem(this.flatNodeMap.get(this.dragNode));
+                this.treeControl.expandDescendants(this.nestedNodeMap.get(newItem));
+            }
         }
         this.dragNode = null;
         this.dragNodeExpandOverNode = null;
         this.dragNodeExpandOverTime = 0;
     }
 
-    public handleDragEnd(event) {
+    public handleDragEnd(event: DragEvent) {
         this.dragNode = null;
         this.dragNodeExpandOverNode = null;
         this.dragNodeExpandOverTime = 0;
